server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.36
httpx==0.27.2
python-dotenv==1.0.1
pydantic-settings==2.5.2
pydantic[email]==2.9.2
slowapi==0.1.9
